Organizing parsers file into a module.
