+++ src/dug/parsers/dbgap_parser.py
@@ -1,0 +1,47 @@
+import logging
+from typing import List
+from xml.etree import ElementTree as ET
+
+from dug import utils as utils
+from ._base import DugElement, FileParser, Indexable, InputFile
+
+logger = logging.getLogger('dug')
+
+
+class DbGaPParser(FileParser):
+    # Class for parsers DBGaP Data dictionary into a set of Dug Elements
+
+    def __call__(self, input_file: InputFile) -> List[Indexable]:
+        logger.debug(input_file)
+        tree = ET.parse(input_file)
+        root = tree.getroot()
+        study_id = root.attrib['study_id']
+        participant_set = root.attrib['participant_set']
+
+        # Parse study name from filehandle
+        study_name = utils.parse_study_name_from_filename(input_file)
+
+        if study_name is None:
+            err_msg = f"Unable to parse DbGaP study name from data dictionary: {input_file}!"
+            logger.error(err_msg)
+            raise IOError(err_msg)
+
+        elements = []
+        for variable in root.iter('variable'):
+            elem = DugElement(elem_id=f"{variable.attrib['id']}.p{participant_set}",
+                              name=variable.find('name').text,
+                              desc=variable.find('description').text.lower(),
+                              elem_type="DbGaP",
+                              collection_id=f"{study_id}.p{participant_set}",
+                              collection_name=study_name)
+
+            # Create DBGaP links as study/variable actions
+            elem.collection_action = utils.get_dbgap_study_link(study_id=elem.collection_id)
+            elem.action = utils.get_dbgap_var_link(study_id=elem.collection_id,
+                                                   variable_id=elem.id.split(".")[0].split("phv")[1])
+            # Add to set of variables
+            logger.debug(elem)
+            elements.append(elem)
+
+        # You don't actually create any concepts
+        return elements
